+++ astro.config.mjs
@@ -34,6 +34,7 @@ export default defineConfig({
       alias: {
         '@': path.resolve(__dirname, './src'),
         '@components': path.resolve(__dirname, './src/components'),
+        '@design-system': path.resolve(__dirname, './src/design-system'),
         '@layouts': path.resolve(__dirname, './src/layouts'),
         '@stores': path.resolve(__dirname, './src/stores'),
         '@types': path.resolve(__dirname, './src/types'),
