+++ tsconfig.json
@@ -5,6 +5,7 @@
     "paths": {
       "@/*": ["src/*"],
       "@components/*": ["src/components/*"],
+      "@design-system/*": ["src/design-system/*"],
       "@layouts/*": ["src/layouts/*"],
       "@stores/*": ["src/stores/*"],
       "@types/*": ["src/types/*"],
